Added discord scraper, refactored some of the folder structure
