--- examples/tweets/tweet_fetcher_config.yaml
@@ -1,25 +1,0 @@
-# API settings
-api_endpoint: 'http://localhost:8080/api/v1/data/twitter/tweets/recent'
-headers:
-  accept: 'application/json'
-  Content-Type: 'application/json'
-
-# Query settings
-query: '(from:macnbtc)'
-tweets_per_request: 200
-
-# Add start_date and end_date
-start_date: '2023-07-24'
-end_date: '2024-07-24'
-days_per_iteration: 10
-
-# File settings
-data_directory: 'data'
-
-# Timing settings
-retry_delay: 960  # 16 minutes in seconds
-request_delay: 15  # 15 seconds between requests
-
-# Logging settings
-log_level: 'INFO'
-log_format: '%(asctime)s - %(levelname)s - %(message)s'
