--- last_known_state.json
@@ -1,1 +1,0 @@
-{"current_date": "2024-07-24"}
