+++ scrapers/tweets/tweet_fetcher_config.yaml
@@ -1,0 +1,25 @@
+# API settings
+api_endpoint: 'http://localhost:8080/api/v1/data/twitter/tweets/recent'
+headers:
+  accept: 'application/json'
+  Content-Type: 'application/json'
+
+# Query settings
+query: '(from:MartiniGuyYT)'
+tweets_per_request: 350
+
+# Add start_date and end_date
+start_date: '2024-07-31'
+end_date: '2024-07-31'
+days_per_iteration: 3
+
+# File settings
+data_directory: 'data/twitter_data'
+
+# Timing settings
+retry_delay: 960  # 16 minutes in seconds
+request_delay: 15  # 15 seconds between requests
+
+# Logging settings
+log_level: 'INFO'
+log_format: '%(asctime)s - %(levelname)s - %(message)s'
